Use `go install` instead of `go get` + updates + remove caching
The build is failing because `go get` is no longer supported outside of
a Go module, which is okay, since what we were really trying to do here
was to do a `go install` instead.

Upgrade `setup-go` and `checkout` actions to v3.

Remove manual caching -- `setup-go` now caches Go dependencies
automatically so it's not doing much for us anymore.